rename package to hchi
