+++ services/horizon/internal/render/problem/main.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"net/http"
 
-	"github.com/stellar/go/support/context/requestid"
+	"github.com/stellar/go/services/horizon/internal/hchi"
 	"github.com/stellar/go/support/render/problem"
 )
 
@@ -12,7 +12,7 @@ import (
 // At present it adds the request's id as the problem's Instance, if available.
 func Inflate(ctx context.Context, p *problem.P) {
 	problem.Inflate(p)
-	p.Instance = requestid.RequestID(ctx)
+	p.Instance = hchi.RequestID(ctx)
 }
 
 // Well-known and reused problems below:
